requirements.txt
discord_py_interactions==5.9.2
python-dotenv==1.0.0
discord~=2.3.2
